requirements.txt
streamlit
torch
numpy
facenet-pytorch
pillow
opencv-python
facenet-pytorch
dlib